% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Log-likelihood of Bernoulli Distribution}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A numeric vector containing binary data (0s and 1s).}
}
\value{
The MLE of the probability parameter.
}
\description{
This function calculates the maximum likelihood estimate of probability
parameter for a Bernoulli distribution, given a vector of binary data.
}
\examples{
data <- c(1, 0, 0, 0, 1, 1, 1)
p_max <- logLikBernoulli(data)
}
