% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Calculate Minimum Sample Size for Desired Power}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{Numeric vector, data for one-sample or first group in two-sample t-test.}

\item{x2}{Optional numeric vector, data for second group in two-sample t-test.}
}
\value{
The minimum number of observations needed.
}
\description{
This function calculates the minimum sample size needed to achieve a desired
power for a one-sample or two-sample t-test, given the input data.
}
\examples{
x1 <- c(2.3, 1.9, 2.1, 2.4, 2.2)
min_n <- minimumN(x1)
print(min_n)
x2 <- c(1.8, 2.0, 1.9, 2.2, 2.1)
min_n <- minimumN(x1, x2)
print(min_n)
}
