% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize Names of Data Frame Columns}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A tibble with standardized column names.
}
\description{
This function standardizes the names of a data frame or tibble by converting
them into small camel case format after replacing any non-alphanumeric characters with underscores.
}
\examples{
data <- data.frame(
  "First Name" = c("John", "Jane"),
  "Last_Name" = c("Doe", "Smith"),
  "Age." = c(25, 30)
)
standardized_data <- standardizeNames(data)
print(standardized_data)
}
