% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCurv.R
\name{survCurv}
\alias{survCurv}
\title{Survival Curve Estimator}
\usage{
survCurv(status, time)
}
\arguments{
\item{status}{Numeric vector indicating the event status.}

\item{time}{Numeric vector of times at which events or censoring occurred.}
}
\value{
Plots the survival curve.
}
\description{
This function calculates and plots the survival curve for right-censored survival data.
It requires a status indicator (1 if the event of interest occurred, 0 if censored) and
the corresponding times.
}
\examples{
# Data is assumed to be loaded and preprocessed outside this function
survCurv(status, time)
}
