% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale a Scaled Vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A scaled numeric vector.}
}
\value{
Numeric vector with the scaling reversed.
}
\description{
This function reverses the scaling transformation applied to a numeric vector.
It restores the vector to its original scale if it has been centered and/or scaled.
}
\examples{
x <- c(1, 2, 3, 4, 5)
scaled_x <- scale(x)
unscaled_x <- unscale(scaled_x)
print(unscaled_x)
}
